Base application

+++ src/main/java/com/efo/controllers/CustomerController.java
@@ -86,7 +86,7 @@ public class CustomerController {
 	}
 	
 	@RequestMapping("editcustomer")
-	public String editCustomer(@ModelAttribute("user_id") int user_id, Model model) {
+	public String editCustomer(@ModelAttribute("user_id") Long user_id, Model model) {
 		User user = userService.retrieve(user_id);
 		user.setRoleString(roleUtils.roleToString(user.getRoles()));
