+++ src/main/java/com/efo/controllers/EmployeeController.java
@@ -143,7 +143,7 @@ public class EmployeeController {
 	}
 	
 	@RequestMapping("editemployee")
-	public String editEmployee(@ModelAttribute("user_id") int user_id, Model model) {
+	public String editEmployee(@ModelAttribute("user_id") Long user_id, Model model) {
 		User user = userService.retrieve(user_id);
 		
 		user.setRoleString(roleUtils.roleToString(user.getRoles()));
