+++ src/main/java/com/efo/dao/EmployeeDao.java
@@ -36,7 +36,7 @@ public class EmployeeDao implements IEmployee {
 	}
 
 	@Override
-	public Employee retrieve(int user_id) {
+	public Employee retrieve(Long user_id) {
 		Session session = session();
 		Employee employee = (Employee) session.createCriteria(Employee.class).add(Restrictions.idEq(user_id)).uniqueResult();
 		session.disconnect();
@@ -53,7 +53,7 @@ public class EmployeeDao implements IEmployee {
 	}
 
 	@Override
-	public void delete(int user_id) {
+	public void delete(Long user_id) {
 		Employee employee = retrieve(user_id);
 		Session session = session();
 		Transaction tx = session.beginTransaction();
