+++ src/main/java/com/efo/entity/Customer.java
@@ -20,7 +20,7 @@ public class Customer implements Serializable{
 	@GenericGenerator(name = "generator", strategy = "foreign", parameters = @Parameter(name = "property", value = "user"))
 	@Id
 	@GeneratedValue(generator = "generator")
-	private Integer user_id;
+	private Long user_id;
 	private String salutation;
 	private String firstname;
 	private String lastname;
@@ -31,11 +31,11 @@ public class Customer implements Serializable{
 	@PrimaryKeyJoinColumn
 	private User user;
 
-	public Integer getUser_id() {
+	public Long getUser_id() {
 		return user_id;
 	}
 
-	public void setUser_id(Integer user_id) {
+	public void setUser_id(Long user_id) {
 		this.user_id = user_id;
 	}
