+++ src/main/java/com/efo/entity/Employee.java
@@ -21,7 +21,7 @@ public class Employee implements Serializable {
 	@GenericGenerator(name = "generator", strategy = "foreign", parameters = @Parameter(name = "property", value = "user"))
 	@Id
 	@GeneratedValue(generator = "generator")
-	private Integer user_id;
+	private Long user_id;
 	
 	private String salutation;
 	private String firstname;
@@ -60,11 +60,11 @@ public class Employee implements Serializable {
 	@PrimaryKeyJoinColumn
 	private User user;
 	
-	public Integer getUser_id() {
+	public Long getUser_id() {
 		return user_id;
 	}
 
-	public void setUser_id(Integer user_id) {
+	public void setUser_id(Long user_id) {
 		this.user_id = user_id;
 	}
