+++ src/main/java/com/efo/interfaces/ICustomer.java
@@ -4,7 +4,7 @@ import com.efo.entity.Customer;
 
 public interface ICustomer {
 	public void create(Customer customer);
-	public Customer retrieve(int user_id);
+	public Customer retrieve(Long user_id);
 	public void update(Customer customer);
-	public void delete(int user_id);
+	public void delete(Long user_id);
 }
