+++ src/main/java/com/efo/interfaces/IEmpFinancial.java
@@ -4,7 +4,7 @@ import com.efo.entity.EmpFinancial;
 
 public interface IEmpFinancial {
 	public void Create(EmpFinancial employee);
-	public EmpFinancial retrieve(int user_id);
+	public EmpFinancial retrieve(Long user_id);
 	public void update(EmpFinancial employee);
 	public void delete(EmpFinancial employee);
 }
