+++ src/main/java/com/efo/interfaces/IEmployee.java
@@ -4,7 +4,7 @@ import com.efo.entity.Employee;
 
 public interface IEmployee {
 	public void create(Employee employee);
-	public Employee retrieve(int user_id);
+	public Employee retrieve(Long user_id);
 	public void update(Employee employee);
-	public void delete(int user_id);
+	public void delete(Long user_id);
 }
