+++ src/main/java/com/efo/interfaces/IUser.java
@@ -4,9 +4,9 @@ import com.efo.entity.User;
 
 public interface IUser {
 	public void create(User userProfile);
-	public User retrieve(int user_id);
+	public User retrieve(Long user_id);
 	public User retrieve(String username);
 	public void update(User userProfile);
 	public void delete(User userProfile);
-	public void delete(int user_id);
+	public void delete(Long user_id);
 }
