+++ src/main/java/com/efo/interfaces/IVendor.java
@@ -4,7 +4,7 @@ import com.efo.entity.Vendor;
 
 public interface IVendor {
 	public void create(Vendor vendor);
-	public Vendor retrieve(int user_id);
+	public Vendor retrieve(Long user_id);
 	public void update(Vendor vendor);
-	public void delete(int user_id);
+	public void delete(Long user_id);
 }
