+++ src/main/java/com/efo/service/CustomerService.java
@@ -27,7 +27,7 @@ public class CustomerService implements ICustomer {
 	}
 
 	@Override
-	public Customer retrieve(int user_id) {
+	public Customer retrieve(Long user_id) {
 		return customerDao.retrieve(user_id);
 	}
 
@@ -37,7 +37,7 @@ public class CustomerService implements ICustomer {
 	}
 
 	@Override
-	public void delete(int user_id) {
+	public void delete(Long user_id) {
 		customerDao.delete(user_id);		
 	}
